# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estudent",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/estudent/api",
            "description": "Rest API Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Common Routes"
                ],
                "summary": "User login",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "indexNum",
                                    "password"
                                ],
                                "properties": {
                                    "indexNum": {
                                        "type": "string",
                                        "example": "2023/2222"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Login Successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Common Routes"
                ],
                "summary": "Logouts user, by deleting its token",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Success: You have logged out and the token was deleted"
                    },
                    "400": {
                        "description": "Error: You are not currently logged in"
                    }
                }
            }
        },
        "/course-exams": {
            "get": {
                "tags": [
                    "Admin Routes"
                ],
                "summary": "Get paginated course exams with filters",
                "operationId": "getCourseExams",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseExamFiltersDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paginated list of course exams",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "content": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CourseExam"
                                            }
                                        },
                                        "total-pages": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "total-elements": {
                                            "type": "integer",
                                            "example": 47
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/course-exams/remaining-course-exams": {
            "get": {
                "tags": [
                    "Common Routes"
                ],
                "summary": "Get all remaining course exams for currently logged-in student in provided exam period",
                "operationId": "c6bfec0fb059d1dd0d07c7133b89cb5f",
                "parameters": [
                    {
                        "name": "for-exam-period-id",
                        "in": "query",
                        "description": "Exam period ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Remaining CourseExams retrieved successfully"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/course-exams/registerable-course-exams": {
            "get": {
                "tags": [
                    "Common Routes"
                ],
                "summary": "Get all course-exams for current user that he did not register for yet in any active exam period",
                "operationId": "a0f0c0beddd775b1ec66bcdfbdbc60ff",
                "responses": {
                    "200": {
                        "description": "Registerable CourseExams retrieved successfully"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/course-exam-reports/{courseExamId}": {
            "get": {
                "tags": [
                    "Admin Routes"
                ],
                "summary": "Download Excel report for a course exam identified by courseExamId",
                "operationId": "getCourseExamReport",
                "parameters": [
                    {
                        "name": "course-exam-id",
                        "in": "path",
                        "description": "ID of the CourseExam to generate the report for, will include all exam registrations for that course exam",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file download",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CourseExamFiltersDTO": {
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "page-size": {
                        "type": "integer",
                        "example": 10
                    },
                    "course-name": {
                        "type": "string",
                        "example": "matematika"
                    },
                    "date-from": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-01"
                    },
                    "date-to": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-30"
                    }
                },
                "type": "object"
            },
...
            "CourseExam": {
                "title": "CourseExam",
                "properties": {
                    "examPeriod": {
                        "$ref": "#/components/schemas/ExamPeriod"
                    },
                    "course": {
                        "$ref": "#/components/schemas/CourseInstance"
                    },
                    "examDateTime": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-20T10:00:00"
                    },
                    "hall": {
                        "type": "string",
                        "example": "A1"
                    }
                },
                "type": "object"
            },
            "CourseInstance": {
                "title": "CourseInstance",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "name": {
                        "type": "string",
                        "example": "Matematika"
                    },
                    "semester": {
                        "type": "integer",
                        "example": 2
                    },
                    "espb": {
                        "type": "integer",
                        "example": 6
                    },
                    "participants": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "ExamPeriod": {
                "title": "ExamPeriod",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "June 2025"
                    },
                    "dateRegisterStart": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-01"
                    },
                    "dateRegisterEnd": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-31"
                    },
                    "dateStart": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-15"
                    },
                    "dateEnd": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-20"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Bearer Token",
                "in": "header",
                "scheme": "https",
                "name": "Authorization"
            }
        }
    },
...
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ],
            "passport": []
        }
    ]
}